Properly handle export when things go wrong.
If the user has no permission or the resource doesn't exist, return
a 404.

Note: Although the `export` view previously appeared to have no
permissions checking, it did because the `get_resource` API function
would raise a 403 response for bad permissions.

This commit improves the codebase by moving HTTP response raising from
the API to the views, as is proper.

closes #172