Fixes indexing so it's possible to search on content_xml.
The `model_attr` kwarg is ignored when `use_template` is used. If using a
template, all fields must be specified in the template.

Upon review, it seems that the complication of using a template was unnecessary,
because the template was so simple that having it separated the indexing
definitions unnecessarily into multiple locations.

Moves definition of `content_xml` into search_indexes.py with the definition
for all other fields.

Adds `content_xml` to the fields being indexed.

Removes repository name and ID from indexing (never should have been there).