Fixed `KeyError` on repository page if search index is reset.
This wouldn't happen in production, but it can happen (for a brief window)
running the application in Docker containers.

Closes #283.

Note about tests:
The system behaves differently during testing than when running. When testing,
a test for this passes without the fix. When Django is running normally,
it will return the `KeyError` after a forced index wipe, but only until
Django is restarted. Then it works normally.

Digging into exactly what changes in Haystack (or Elasticsearch) is too
time-consuming for the payback. As a result, this branch causes a minor
hit in test coverage.