Fixed display of vocabulary terms containing spaces.
fixes #367

I spent some (too much) time on the command line using the elasticsearch
module trying to figure out how to fix the search to work with spaces.

In the end, I went for the obvious, lame solution. Eventually it seems
like we'll want to ditch Haystack and use Elasticsearch directly.