Avoided hitting the database for the search page.
Load data elements from LearningResources into the Elasticsearch index,
so they can be inserted into the template from the query results, not
requiring a DB lookup.

Also added a test which should catch regressions. There are still queries
due to the checks to authentication and session management, but there
are no queries to learningresources_learningresource.